+++ src/components/amos-chat/bubble/cloud.sc.js
@@ -1,5 +1,5 @@
 import styled, {css} from 'styled-components'
-import {AVATAR_SIZE_REGULAR, AVATAR_SIZE_LARGE} from 'components/amos-monolog/constants'
+import {AVATAR_SIZE_REGULAR, AVATAR_SIZE_LARGE} from '../constants'
 
 const regular = css`
   /* 21.7 is half of height of first line of bubble */
