+++ src/components/amos-chat/bubble/index.js
@@ -2,14 +2,11 @@ import Cloud from './cloud.sc'
 import React from 'react'
 import Tail from './tail.sc'
 
-const Bubble = ({size, children}) => {
-  console.log('size', size)
-  return (
+const Bubble = ({size, children}) => (
   <Cloud size={size}>
     <Tail/>
     {children}
   </Cloud>
-  )
-}
+)
 
 export default Bubble
