+++ src/containers/sign-up/index.js
@@ -1,12 +1,12 @@
 import Email from './email'
 import React from 'react'
 import Success from './success'
-import Welcome from './welcome'
+import Guest from './guest'
 import {Switch, Route} from 'react-router-dom'
 
 const SignUp = () => (
   <Switch>
-    <Route path="/sign-up" component={Welcome} exact/>
+    <Route path="/sign-up" component={Guest} exact/>
     <Route path="/sign-up/email" component={Email}/>
     <Route path="/sign-up/success" component={Success}/>
   </Switch>
